pyproject.toml
[project]
name = "precord"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "cryptography>=43.0.3",
    "fastapi[standard]==0.115.5",
    "gunicorn>=23.0.0",
    "httpx>=0.27.2",
    "jinja2>=3.1.4",
    "psycopg>=3.2.3",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.1",
    "pyjwt>=2.10.0",
    "rich>=13.9.4",
    "svcs>=24.1.0",
    "typer>=0.13.0",
    "uvicorn>=0.32.0",
]

[dependency-groups]
dev = ["asyncpg-stubs>=0.30.0", "black>=24.10.0", "mypy>=1.13.0", "ruff>=0.7.4"]

[tool.black]
line-length = 98
target-version = ['py312']

[tool.mypy]
strict = true

[tool.ruff]
line-length = 98
target-version = "py312"

[tool.ruff.lint]
select = ["ALL"]
ignore = ["ANN401", "S101"]
fixable = ["ALL"]

[tool.ruff.lint.isort]
combine-as-imports = true
